+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Mongo.FileStorage
 
+## Release 8.1.5
+
+- Add method `DownloadAsStreamAsync`.
+- Add method `DownloadAsByteArrayAsync`.
+- Add method `GetFileInfoAsync`.
+- Mark method `DownloadAsync` as Obsolete
+- Mark method `GetAsync` as Obsolete
+
 ## Release 8.1.4
 
 - Minor fixes